requirements.txt
fastapi
uvicorn
motor
python-dotenv
pydantic[dotenv]
pymongo
passlib[bcrypt]
pyjwt
python-multipart
PyPDF2
python-magic==0.4.27
python-jose[cryptography]
celery
redis
httpx==0.28.0
pytest
pytest-asyncio
pydantic-settings==2.2.1
pydantic[email]



google-generativeai
fastapi-mail
pyzoom
openai


google-api-python-client
google-auth
google-auth-oauthlib